+++ Database/his/Tables/MatchDetails.sql
@@ -63,5 +63,6 @@ CREATE TABLE [his].[matchdetails](
 	[UitTeamEmail] [nvarchar](200) NULL,
 	[mta_inserted] [datetime] NULL,
 	[mta_modified] [datetime] NULL,
-	[mta_deleted] [datetime] NULL
+	[mta_deleted] [datetime] NULL,
+	[ClubCode]     [nvarchar](20) NULL
 )